+++ lib/global_id_utils.rb
@@ -2,11 +2,10 @@ require 'active_support'
 require 'active_model'
 require 'globalid'
 
-require 'global_id_utils/locator'
-require 'global_id_model'
 require 'global_id_validator'
+require 'global_id_utils/locator'
+require 'global_id_utils/model_extensions'
 require 'global_id_utils/railtie'
-
 
 class GlobalID
   module Locator
