+++ lib/global_id_utils/railtie.rb
@@ -2,7 +2,7 @@ module GlobalIdUtils
   class Railtie < Rails::Railtie
     initializer 'global_id_utils' do |app|
       ActiveSupport.on_load(:active_record) do
-        include GlobalIdModel
+        include ::GlobalIdUtils::ModelExtensions
       end
     end
   end
